Made data user specific

+++ InventoryApp/Models/DataManager.swift
@@ -21,28 +21,36 @@ class DataManager: ObservableObject{
 //    
     func fetchItems(){
         inventory.removeAll()
-        let db = Firestore.firestore()
-        let ref = db.collection("Items")
-        ref.getDocuments { snapshot, error in
-            guard error == nil else{
-                print(error!.localizedDescription)
-                return
-            }
-            
-            if let snapshot = snapshot{
-                for document in snapshot.documents{
-                    let data = document.data()
-                    let notes = data["notes"] as? String ?? ""
-                    let name = data["name"] as? String ?? ""
-                    let amountInStock = data["amountInStock"] as? Int ?? 0
-                    let amountTotal = data["amountTotal"] as? Int ?? 0
-                    let item = Item(name: name, notes:notes, amountTotal: amountTotal, amountInStock: amountInStock)
-                    self.inventory.append(item)
-                }
-            }
+        if let currentUser = Auth.auth().currentUser{
+            let userID = currentUser.uid
+            let db = Firestore.firestore()
+            let fullPath = "Users/\(userID)/Items"
+            let ref = db.collection(fullPath)
+                  ref.getDocuments { snapshot, error in
+                      guard error == nil else{
+                          print(error!.localizedDescription)
+                          return
+                      }
+                      
+                      if let snapshot = snapshot{
+                          for document in snapshot.documents{
+                              let data = document.data()
+                              let amountInStock = data["amountInStock"] as? Int ?? 50
+                              let amountTotal = data["amountTotal"] as? Int ?? 50
+                              let name = data["name"] as? String ?? ""
+                              let notes = data["notes"] as? String ?? ""
+                           
+                            
+                              let item = Item(name: name, notes: notes, amountTotal: amountTotal, amountInStock: amountInStock)
+                              self.inventory.append(item)
+                            
+                          }
+                      }
+                  }
+                  print("Fetch request")
+                  hasLoadedData = true
         }
-        print("Fetch request")
-        hasLoadedData = true
+        
     }
     
     
@@ -50,44 +58,57 @@ class DataManager: ObservableObject{
         return inventory
     }
     
-    func addItem(itemName: String, itemNotes: String, itemTotal: String){
-        let db = Firestore.firestore()
-        let ref = db.collection("Items").document(itemName)
-        ref.setData(["name":itemName, "notes":itemNotes, "amountTotal":Int(itemTotal), "amountInStock":Int(itemTotal)]){ error in
-            if let error = error{
-                print(error.localizedDescription)
+    func addItem(itemName: String, itemNotes: String, itemAmount: String){
+        if let currentUser = Auth.auth().currentUser{
+            let userID = currentUser.uid
+            let db = Firestore.firestore()
+            let fullPath = "Users/\(userID)/Items/\(itemName)"
+            let ref = db.document(fullPath)
+    
+          
+            ref.setData(["name": itemName, "notes": itemNotes, "amountTotal": Int(itemAmount), "amountInStock": Int(itemAmount)]){ error in
+                if let error = error{
+                    print(error.localizedDescription)
+                }
+                
             }
-            
         }
+       
   
     }
     
     func updateItem(itemName: String, itemStock: Int){
-        let db = Firestore.firestore()
-        let ref = db.collection("Items").document(itemName)
-        ref.updateData(["amountInStock":itemStock]){ error in
-            if let error = error{
-                print(error.localizedDescription)
+        if let currentUser = Auth.auth().currentUser{
+            let userID = currentUser.uid
+            let db = Firestore.firestore()
+            let fullPath = "Users/\(userID)/Items/\(itemName)"
+            let ref = db.document(fullPath)
+            ref.updateData(["amountInStock":itemStock]){ error in
+                if let error = error{
+                    print(error.localizedDescription)
+                }
+                
             }
-            
+            fetchItems()
         }
-        fetchItems()
+       
        
     }
     
     func deleteItem(itemName: String){
-        let db = Firestore.firestore()
-        let ref = db.collection("Items").document(itemName)
-        ref.delete(){ error in
-            if let error = error{
-                print(error.localizedDescription)
+        if let currentUser = Auth.auth().currentUser{
+            let userID = currentUser.uid
+            let db = Firestore.firestore()
+            let fullPath = "Users/\(userID)/Items/\(itemName)"
+            let ref = db.document(fullPath)
+            ref.delete(){ error in
+                if let error = error{
+                    print(error.localizedDescription)
+                }
+                
             }
-            
+            fetchItems()
         }
-        fetchItems()
-   
-        
-
     }
